Feature umar api integration

+++ src/App.js
@@ -10,13 +10,15 @@ function App() {
   const check = localStorage.getItem("check");
   useEffect(() => {
     window.addEventListener("beforeunload", function (e) {
-      if (check === "false") {
+      if (
+        e.currentTarget.performance.navigation.type !==
+          PerformanceNavigation.TYPE_RELOAD &&
+        check === "false"
+      ) {
         localStorage.clear();
       }
     });
   }, []);
-
-  // e.currentTarget.performance.navigation.type === PerformanceNavigation.TYPE_RELOAD
 
   return (
     <>
